✨ Add companies service

+++ companies/src/main/kotlin/org/seasonist/CompanyResource.kt
@@ -2,7 +2,6 @@ package org.seasonist
 
 import jakarta.enterprise.context.ApplicationScoped
 import jakarta.transaction.Transactional
-import org.eclipse.microprofile.graphql.DefaultValue
 import org.eclipse.microprofile.graphql.Description
 import org.eclipse.microprofile.graphql.GraphQLApi
 import org.eclipse.microprofile.graphql.Mutation
@@ -15,7 +14,10 @@ class CompanyResource(val repository: CompanyRepository) {
 
     @Query
     @Description("Find a company by id")
-    fun findById(id: String): Company = repository.findById(UUID.fromString(id))
+    fun findById(id: String): Company? {
+        try { UUID.fromString(id) } catch (e: IllegalArgumentException) { return null }
+        return repository.findById(UUID.fromString(id))
+    }
 
     @Query
     @Description("List all companies")
@@ -33,20 +35,28 @@ class CompanyResource(val repository: CompanyRepository) {
     @Mutation
     @Transactional
     @Description("Update company")
-    fun updateCompany(id: String, name: String, address: String): Company {
+    fun updateCompany(id: String, name: String, address: String): Company? {
+        try { UUID.fromString(id) } catch (e: IllegalArgumentException) { return null }
         val company = repository.findById(UUID.fromString(id))
-        company.name = name
-        company.address = address
-        repository.persist(company)
-        return company
+        if (company != null) {
+            company.name = name
+            company.address = address
+            repository.persist(company)
+            return company
+        }
+        return null
     }
 
     @Mutation
     @Transactional
     @Description("Delete company")
-    fun deleteCompany(id: String): Company {
+    fun deleteCompany(id: String): Company? {
+        try { UUID.fromString(id) } catch (e: IllegalArgumentException) { return null }
         val company = repository.findById(UUID.fromString(id))
-        repository.delete(company)
-        return company
+        if (company != null) {
+            repository.delete(company)
+            return company
+        }
+        return null
     }
 }
